Increase precision of frame rate property
Summary:
Some video files have frame rates
of 23.976 or similar, requiring the representation
as double. The data will be displayed
with two digit precision, as done by
kdenlive for example.

Reviewers: bruns, ngraham

Reviewed By: bruns

Subscribers: kde-frameworks-devel, #baloo

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D19101

+++ src/propertyinfo.cpp
@@ -143,7 +143,7 @@ PropertyInfo::PropertyInfo(Property::Property property)
         case Property::FrameRate:
             d->name = QStringLiteral("frameRate");
             d->displayName = i18nc("@label", "Frame Rate");
-            d->valueType = QVariant::Int;
+            d->valueType = QVariant::Double;
             d->formatAsString = &FormatStrings::formatAsFrameRate;
             break;
